tracee-rules: extract getSignatures
Extract getSignatures to a new package, and make it public func.
