+++ pkg/rules/signature/signature.go
@@ -1,4 +1,4 @@
-package main
+package signature
 
 import (
 	_ "embed"
@@ -20,7 +20,7 @@ import (
 	"github.com/aquasecurity/tracee/types/detect"
 )
 
-func getSignatures(target string, partialEval bool, rulesDir string, rules []string, aioEnabled bool) ([]detect.Signature, error) {
+func Find(target string, partialEval bool, rulesDir string, rules []string, aioEnabled bool) ([]detect.Signature, error) {
 	if rulesDir == "" {
 		exePath, err := os.Executable()
 		if err != nil {
